+++ src/Services/CRM/Common/Result/AbstractCrmItem.php
@@ -51,6 +51,7 @@ class AbstractCrmItem extends AbstractItem
 
         switch ($offset) {
             case 'ID':
+            case 'NOTIFY_VALUE':
             case 'ASSIGNED_BY_ID':
             case 'CONTACT_IDS':
             case 'RESPONSIBLE_ID':
@@ -85,6 +86,7 @@ class AbstractCrmItem extends AbstractItem
             case 'RESULT_STREAM':
             case 'LAST_ACTIVITY_BY':
             case 'ADDRESS_LOC_ADDR_ID':
+            case 'OWNER_TYPE_ID':
                 if ($this->data[$offset] !== '' && $this->data[$offset] !== null) {
                     return (int)$this->data[$offset];
                 }
@@ -114,6 +116,7 @@ class AbstractCrmItem extends AbstractItem
             case 'TAX_INCLUDED':
             case 'CUSTOMIZED':
             case 'COMPLETED':
+            case 'IS_INCOMING_CHANNEL':
                 return $this->data[$offset] === 'Y';
             case 'DATE_CREATE':
             case 'CREATED_DATE':
@@ -129,6 +132,8 @@ class AbstractCrmItem extends AbstractItem
             case 'movedTime':
             case 'lastActivityTime':
             case 'LAST_ACTIVITY_TIME':
+            case 'START_TIME':
+            case 'END_TIME':
                 if ($this->data[$offset] !== '') {
                     return CarbonImmutable::createFromFormat(DATE_ATOM, $this->data[$offset]);
                 }
@@ -140,6 +145,7 @@ class AbstractCrmItem extends AbstractItem
             case 'PRICE':
             case 'DISCOUNT_SUM':
             case 'RESULT_SUM':
+            case 'RESULT_VALUE':
                 if ($this->data[$offset] !== '' && $this->data[$offset] !== null) {
                     $var = $this->data[$offset] * 100;
                     return new Money((string)$var, new Currency($this->currency->getCode()));