+++ tests/CustomAssertions/CustomBitrix24Assertions.php
@@ -12,8 +12,16 @@ declare(strict_types=1);
 
 namespace Bitrix24\SDK\Tests\CustomAssertions;
 
-use Bitrix24\SDK\Services\CRM\Contact\Result\ContactItemResult;
+use Bitrix24\SDK\Services\CRM\Activity\ActivityContentType;
+use Bitrix24\SDK\Services\CRM\Activity\ActivityDirectionType;
+use Bitrix24\SDK\Services\CRM\Activity\ActivityNotifyType;
+use Bitrix24\SDK\Services\CRM\Activity\ActivityPriority;
+use Bitrix24\SDK\Services\CRM\Activity\ActivityStatus;
+use Bitrix24\SDK\Services\CRM\Activity\ActivityType;
+use Carbon\CarbonImmutable;
 use Typhoon\Reflection\TyphoonReflector;
+use function Typhoon\Type\stringify;
+use Money\Currency;
 
 trait CustomBitrix24Assertions
 {
@@ -41,5 +49,236 @@ trait CustomBitrix24Assertions
                 $resultItemClassName,
                 implode(', ', array_values(array_diff($fieldCodesFromApi, $propsFromAnnotations)))
             ));
+    }
+
+    protected function assertBitrix24AllResultItemFieldsHasValidTypeAnnotation(array $fieldCodesFromApi, string $resultItemClassName):void
+    {
+        // parse keys from phpdoc annotation
+        $props = TyphoonReflector::build()->reflectClass($resultItemClassName)->properties();
+        $propsFromAnnotations = [];
+        foreach ($props as $meta) {
+            if ($meta->isAnnotated() && !$meta->isNative()) {
+                $propsFromAnnotations[$meta->id->name]=stringify($meta->type());
+            }
+        }
+
+        asort($propsFromAnnotations);
+        asort($fieldCodesFromApi);
+        foreach($fieldCodesFromApi as $fieldCode => $fieldData){
+            // mapping internal bitrix24 types to bitrix24 sdk types
+            switch ($fieldData['type']){
+                case 'string':
+                case 'crm_status':
+                    // if field code contains currency
+                    if(str_contains($fieldCode, 'CURRENCY_ID')){
+                        $this->assertTrue(
+                            str_contains($propsFromAnnotations[$fieldCode], Currency::class),
+                            sprintf('class «%s» field «%s» has invalid type phpdoc annotation «%s», field type from bitrix24 is «%s», expected sdk-type «%s»',
+                                $resultItemClassName,
+                                $fieldCode,
+                                $propsFromAnnotations[$fieldCode],
+                                $fieldData['type'],
+                                Currency::class
+                            )
+                        );
+                        break;                        
+                    }
+                    
+                    $this->assertTrue(
+                        str_contains($propsFromAnnotations[$fieldCode], 'string',),
+                        sprintf('class «%s» field «%s» has invalid type phpdoc annotation «%s», field type from bitrix24 is «%s», expected sdk-type «%s»',
+                            $resultItemClassName,
+                            $fieldCode,
+                            $propsFromAnnotations[$fieldCode],
+                            $fieldData['type'],
+                            'string'
+                        )
+                    );
+                    break;
+                case 'user':
+                case 'crm_enum_ownertype':
+                case 'crm_lead':
+                case 'integer':
+                    $this->assertTrue(
+                        str_contains($propsFromAnnotations[$fieldCode],'int'),
+                        sprintf('class «%s» field «%s» has invalid type phpdoc annotation «%s», field type from bitrix24 is «%s», expected sdk-type «%s»',
+                            $resultItemClassName,
+                            $fieldCode,
+                            $propsFromAnnotations[$fieldCode],
+                            $fieldData['type'],
+                            'int'
+                        )
+                    );
+                    break;
+                case 'double':
+                    $this->assertEquals(
+                        'Money\Money|null',
+                        $propsFromAnnotations[$fieldCode],
+                        sprintf('class «%s» field «%s» has invalid type phpdoc annotation «%s», field type from bitrix24 is «%s», expected sdk-type «%s»',
+                            $resultItemClassName,
+                            $fieldCode,
+                            $propsFromAnnotations[$fieldCode],
+                            $fieldData['type'],
+                            'Money\Money|null'
+                        )
+                    );
+                    break;
+                case 'date':
+                case 'datetime':
+                    $this->assertTrue(
+                        str_contains($propsFromAnnotations[$fieldCode],CarbonImmutable::class),
+                        sprintf('class «%s» field «%s» has invalid type phpdoc annotation «%s», field type from bitrix24 is «%s», expected sdk-type «%s»',
+                            $resultItemClassName,
+                            $fieldCode,
+                            $propsFromAnnotations[$fieldCode],
+                            $fieldData['type'],
+                            CarbonImmutable::class
+                        )
+                    );
+                    break;
+                case 'char':
+                    $this->assertEquals(
+                        'bool',
+                        $propsFromAnnotations[$fieldCode],
+                        sprintf('class «%s» field «%s» has invalid type phpdoc annotation «%s», field type from bitrix24 is «%s», expected sdk-type «%s»',
+                            $resultItemClassName,
+                            $fieldCode,
+                            $propsFromAnnotations[$fieldCode],
+                            $fieldData['type'],
+                            'bool'
+                        )
+                    );
+                    break;
+                case 'diskfile':
+                case 'object':
+                case 'file':
+                case 'crm_company':
+                    if(str_contains($fieldCode, '_IDS') || str_contains($fieldCode, 'PHOTO')) {
+                        $this->assertTrue(
+                            str_contains($propsFromAnnotations[$fieldCode], 'array'),
+                            sprintf('class «%s» field «%s» has invalid type phpdoc annotation «%s», field type from bitrix24 is «%s», expected sdk-type «%s»',
+                                $resultItemClassName,
+                                $fieldCode,
+                                $propsFromAnnotations[$fieldCode],
+                                $fieldData['type'],
+                                'array'
+                            )
+                        );
+                        break;
+                    }
+                    $this->assertTrue(
+                        str_contains($propsFromAnnotations[$fieldCode], 'int'),
+                        sprintf('class «%s» field «%s» has invalid type phpdoc annotation «%s», field type from bitrix24 is «%s», expected sdk-type «%s»',
+                            $resultItemClassName,
+                            $fieldCode,
+                            $propsFromAnnotations[$fieldCode],
+                            $fieldData['type'],
+                            'int'
+                        )
+                    );
+                    break;
+                case 'crm_enum_activitydirection':
+                    $this->assertEquals(
+                        ActivityDirectionType::class,
+                        $propsFromAnnotations[$fieldCode],
+                        sprintf('class «%s» field «%s» has invalid type phpdoc annotation «%s», field type from bitrix24 is «%s», expected sdk-type «%s»',
+                            $resultItemClassName,
+                            $fieldCode,
+                            $propsFromAnnotations[$fieldCode],
+                            $fieldData['type'],
+                            ActivityDirectionType::class
+                        )
+                    );
+                    break;
+                case 'crm_enum_contenttype':
+                        $this->assertEquals(
+                            ActivityContentType::class,
+                            $propsFromAnnotations[$fieldCode],
+                            sprintf('class «%s» field «%s» has invalid type phpdoc annotation «%s», field type from bitrix24 is «%s», expected sdk-type «%s»',
+                                $resultItemClassName,
+                                $fieldCode,
+                                $propsFromAnnotations[$fieldCode],
+                                $fieldData['type'],
+                                ActivityContentType::class
+                            )
+                        );
+                        break;
+                case 'crm_enum_activitytype':
+                    $this->assertEquals(
+                        ActivityType::class,
+                        $propsFromAnnotations[$fieldCode],
+                        sprintf('class «%s» field «%s» has invalid type phpdoc annotation «%s», field type from bitrix24 is «%s», expected sdk-type «%s»',
+                            $resultItemClassName,
+                            $fieldCode,
+                            $propsFromAnnotations[$fieldCode],
+                            $fieldData['type'],
+                            ActivityType::class
+                        )
+                    );
+                    break;
+                case 'crm_enum_activitynotifytype':
+                    $this->assertEquals(
+                        ActivityNotifyType::class,
+                        $propsFromAnnotations[$fieldCode],
+                        sprintf('class «%s» field «%s» has invalid type phpdoc annotation «%s», field type from bitrix24 is «%s», expected sdk-type «%s»',
+                            $resultItemClassName,
+                            $fieldCode,
+                            $propsFromAnnotations[$fieldCode],
+                            $fieldData['type'],
+                            ActivityNotifyType::class
+                        )
+                    );
+                    break;
+                case 'crm_enum_activitypriority':
+                    $this->assertEquals(
+                        ActivityPriority::class,
+                        $propsFromAnnotations[$fieldCode],
+                        sprintf('class «%s» field «%s» has invalid type phpdoc annotation «%s», field type from bitrix24 is «%s», expected sdk-type «%s»',
+                            $resultItemClassName,
+                            $fieldCode,
+                            $propsFromAnnotations[$fieldCode],
+                            $fieldData['type'],
+                            ActivityPriority::class
+                        )
+                    );
+                    break;
+                case 'crm_enum_activitystatus':
+                    $this->assertEquals(
+                        ActivityStatus::class,
+                        $propsFromAnnotations[$fieldCode],
+                        sprintf('class «%s» field «%s» has invalid type phpdoc annotation «%s», field type from bitrix24 is «%s», expected sdk-type «%s»',
+                            $resultItemClassName,
+                            $fieldCode,
+                            $propsFromAnnotations[$fieldCode],
+                            $fieldData['type'],
+                            ActivityStatus::class
+                        )
+                    );
+                    break;
+                case 'crm_activity_binding':
+                case 'crm_activity_communication':
+                case 'crm_multifield':
+                        $this->assertTrue(
+                            str_contains($propsFromAnnotations[$fieldCode],'array'),
+                            sprintf('class «%s» field «%s» has invalid type phpdoc annotation «%s», field type from bitrix24 is «%s», expected sdk-type «%s»',
+                                $resultItemClassName,
+                                $fieldCode,
+                                $propsFromAnnotations[$fieldCode],
+                                $fieldData['type'],
+                                'array'
+                            )
+                        );
+                    break;
+                default:
+                    $this->assertFalse(true,
+                        sprintf('class «%s» field «%s» has unknown field type from bitrix24 «%s», sdk-type from annotation «%s», fix type mapping map in integration test',
+                            $resultItemClassName,
+                            $fieldCode,
+                            $fieldData['type'],
+                            $propsFromAnnotations[$fieldCode],
+                        )
+                    );
+            }
+        }
     }
 }